Rename 'customers' to 'customerList'
