+++ src/features/customerList/customerDetails/CustomerDetails.tsx
