+++ src/features/customerList/customerDetails/customerDetails.module.scss
